+++ src/Entities/Definitions/InterfaceDefinition.php
@@ -33,11 +33,6 @@ use AppserverIo\Doppelgaenger\Entities\Lists\TypedListList;
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  * @link      https://github.com/appserver-io/doppelgaenger
  * @link      http://www.appserver.io/
- *
- * @property array                                                             $extends              Parental interfaces (if any)
- * @property array                                                             $constants            Possible constants the interface defines
- * @property \AppserverIo\Doppelgaenger\Entities\Lists\AssertionList           $invariantConditions  List of directly defined invariant conditions
- * @property \AppserverIo\Doppelgaenger\Entities\Lists\TypedListList           $ancestralInvariants  List of lists of any ancestral invariants
  */
 class InterfaceDefinition extends AbstractStructureDefinition
 {