+++ src/StreamFilters/EnforcementFilter.php
@@ -20,10 +20,14 @@
 
 namespace AppserverIo\Doppelgaenger\StreamFilters;
 
+use AppserverIo\Doppelgaenger\Config;
+use AppserverIo\Doppelgaenger\Entities\Definitions\FunctionDefinition;
 use AppserverIo\Doppelgaenger\Exceptions\ExceptionFactory;
 use AppserverIo\Doppelgaenger\Exceptions\GeneratorException;
 use AppserverIo\Doppelgaenger\Dictionaries\Placeholders;
 use AppserverIo\Doppelgaenger\Dictionaries\ReservedKeywords;
+use AppserverIo\Doppelgaenger\Interfaces\StructureDefinitionInterface;
+use AppserverIo\Psr\MetaobjectProtocol\Dbc\Annotations\Processing;
 
 /**
  * This filter will buffer the input stream and add the processing information into the prepared assertion checks
@@ -50,7 +54,7 @@ class EnforcementFilter extends AbstractFilter
      *
      * @var array $dependencies
      */
-    protected $dependencies = array('PreconditionFilter', 'PostconditionFilter', 'InvariantFilter');
+    protected $dependencies = array(array('PreconditionFilter', 'PostconditionFilter', 'InvariantFilter'));
 
     /**
      * The main filter method.
@@ -71,31 +75,74 @@ class EnforcementFilter extends AbstractFilter
     public function filter($in, $out, &$consumed, $closing)
     {
         // Lets check if we got the config we wanted
-        $config = $this->params;
+        $config = $this->params['config'];
+        $structureDefinition = $this->params['structureDefinition'];
 
+        // check if we got what we need for proper processing
+        if (!$config instanceof Config || !$structureDefinition instanceof StructureDefinitionInterface) {
+            throw new GeneratorException('The enforcement filter needs the configuration as well as the definition of the currently filtered structure. At least one of these requirements is missing.');
+        }
+
+        // we need a valid configuration as well
         if (!$config->hasValue('enforcement/processing')) {
             throw new GeneratorException('Configuration does not contain the needed processing section.');
         }
 
+        // get the default enforcement processing
+        $localType = $this->filterLocalProcessing($structureDefinition->getDocBlock());
+        $type = $localType ? $localType : $config->getValue('enforcement/processing');
+
         // Get the code for the processing
-        $preconditionCode = $this->generateCode($config, 'precondition');
-        $postconditionCode = $this->generateCode($config, 'postcondition');
-        $invariantCode = $this->generateCode($config, 'invariant');
-        $invalidCode = $this->generateCode($config, 'InvalidArgumentException');
-        $missingCode = $this->generateCode($config, 'MissingPropertyException');
+        $structureName = $structureDefinition->getQualifiedName();
+        $preconditionCode = $this->generateCode($structureName, 'precondition', $type);
+        $postconditionCode = $this->generateCode($structureName, 'postcondition', $type);
+        $invariantCode = $this->generateCode($structureName, 'invariant', $type);
+        $invalidCode = $this->generateCode($structureName, 'InvalidArgumentException', $type);
+        $missingCode = $this->generateCode($structureName, 'MissingPropertyException', $type);
 
         // Get our buckets from the stream
         while ($bucket = stream_bucket_make_writeable($in)) {
+            // Get the tokens
+            $tokens = token_get_all($bucket->data);
+
+            // Go through the tokens and check what we found
+            $tokensCount = count($tokens);
+            for ($i = 0; $i < $tokensCount; $i++) {
+                // Did we find a function? If so check if we know that thing and insert the code of its preconditions.
+                if (is_array($tokens[$i]) && $tokens[$i][0] === T_FUNCTION && is_array($tokens[$i + 2])) {
+                    // Get the name of the function
+                    $functionName = $tokens[$i + 2][1];
+
+                    // Check if we got the function in our list, if not continue
+                    $functionDefinition = $structureDefinition->getFunctionDefinitions()->get($functionName);
+
+                    if (!$functionDefinition instanceof FunctionDefinition) {
+                        continue;
+
+                    } else {
+                        // manage the injection of the enforcement code into the found function
+                        $this->injectFunctionEnforcement(
+                            $bucket->data,
+                            $structureName,
+                            $preconditionCode,
+                            $postconditionCode,
+                            $functionDefinition
+                        );
+
+                        // "Destroy" code and function definition
+                        $functionDefinition = null;
+                    }
+                }
+            }
+
             // Insert the code for the static processing placeholders
             $bucket->data = str_replace(
                 array(
-                    Placeholders::PROCESSING . 'precondition' . Placeholders::PLACEHOLDER_CLOSE,
-                    Placeholders::PROCESSING . 'postcondition' . Placeholders::PLACEHOLDER_CLOSE,
-                    Placeholders::PROCESSING . 'invariant' . Placeholders::PLACEHOLDER_CLOSE,
-                    Placeholders::PROCESSING . 'InvalidArgumentException' . Placeholders::PLACEHOLDER_CLOSE,
-                    Placeholders::PROCESSING . 'MissingPropertyException' . Placeholders::PLACEHOLDER_CLOSE
+                    Placeholders::ENFORCEMENT . 'invariant' . Placeholders::PLACEHOLDER_CLOSE,
+                    Placeholders::ENFORCEMENT . 'InvalidArgumentException' . Placeholders::PLACEHOLDER_CLOSE,
+                    Placeholders::ENFORCEMENT . 'MissingPropertyException' . Placeholders::PLACEHOLDER_CLOSE
                 ),
-                array($preconditionCode, $postconditionCode, $invariantCode, $invalidCode, $missingCode),
+                array($invariantCode, $invalidCode, $missingCode),
                 $bucket->data
             );
 
@@ -108,15 +155,44 @@ class EnforcementFilter extends AbstractFilter
     }
 
     /**
+     * Will try to filter custom local enforcement processing from a given docBloc.
+     * Will return the found value, FALSE otherwise
+     *
+     * @param string $docBlock DocBloc to filter
+     *
+     * @return boolean|string
+     */
+    protected function filterLocalProcessing($docBlock)
+    {
+        // if the annotation cannot be found we have to do nothing here
+        if (strpos($docBlock, '@' . Processing::ANNOTATION) === false) {
+            return false;
+        }
+
+        // try to preg_match the right annotation
+        $matches = array();
+        preg_match_all('/@' . Processing::ANNOTATION . '\("(.+)"\)/', $docBlock, $matches);
+        if (isset($matches[1])) {
+            return array_pop($matches[1]);
+        }
+
+        // still here? Tell them we have failed then
+        return false;
+    }
+
+    /**
      * /**
      * Will generate the code needed to enforce any broken assertion checks
      *
-     * @param \AppserverIo\Doppelgaenger\Config $config The configuration aspect which holds needed information for us
-     * @param string                            $for    For which kind of assertion do wee need the processing
+     * @param string $structureName The name of the structure for which we create the enforcement code
+     * @param string $target        For which kind of assertion do wee need the processing
+     * @param string $type          The enforcement processing type to generate code for
      *
      * @return string
+     *
+     * @throws \AppserverIo\Doppelgaenger\Exceptions\GeneratorException
      */
-    private function generateCode($config, $for)
+    protected function generateCode($structureName, $target, $type)
     {
         $code = '';
 
@@ -124,37 +200,100 @@ class EnforcementFilter extends AbstractFilter
         $place = '__METHOD__';
 
         // If we are in an invariant we should tell them about the method we got called from
-        if ($for === 'invariant') {
+        if ($target === 'invariant') {
             $place = '$callingMethod';
         }
 
-        // What kind of reaction should we create?
-        switch ($config->getValue('enforcement/processing')) {
+        // what we will always need is collection of all errors that occurred
+        $errorCollectionCode = 'if (empty(' . ReservedKeywords::FAILURE_VARIABLE . ')) {
+            ' . ReservedKeywords::FAILURE_VARIABLE . ' = "";
+            } else {
+                ' . ReservedKeywords::FAILURE_VARIABLE . ' = \'Failed ' . $target . ' "\' . implode(\'" and "\', ' . ReservedKeywords::FAILURE_VARIABLE . ') . \'" in \' . ' . $place . ';
+            }' .
+            ReservedKeywords::FAILURE_VARIABLE . ' .= implode(" and ", ' . ReservedKeywords::UNWRAPPED_FAILURE_VARIABLE . ');';
+
+        // what kind of processing should we create?
+        switch ($type) {
             case 'exception':
 
                 $exceptionFactory = new ExceptionFactory();
-                $exception = $exceptionFactory->getClassName($for);
+                $exception = $exceptionFactory->getClassName($target);
 
                 // Create the code
-                $code .= '\AppserverIo\Doppelgaenger\ContractContext::close();
-                throw new \\' . $exception . '("Failed ' . ReservedKeywords::FAILURE_VARIABLE . ' in " . ' . $place . ');';
+                $code .= '\AppserverIo\Doppelgaenger\ContractContext::close();' .
+                    $errorCollectionCode . '
+                    throw new \\' . $exception . '(' . ReservedKeywords::FAILURE_VARIABLE . ');';
 
                 break;
 
             case 'logging':
 
                 // Create the code
-                $code .= '$container = new \AppserverIo\Doppelgaenger\Utils\InstanceContainer();
-                $logger = $container[' . ReservedKeywords::LOGGER_CONTAINER_ENTRY . '];
-                $logger->error("Failed ' . $for .
-                    ReservedKeywords::FAILURE_VARIABLE . ' in " . ' . $place . ');';
+                $code .= $errorCollectionCode .
+                    '$container = new \AppserverIo\Doppelgaenger\Utils\InstanceContainer();
+                    $logger = @$container[\'' . ReservedKeywords::LOGGER_CONTAINER_ENTRY . '\'];
+                    if (is_null($logger)) {
+                        error_log(' . ReservedKeywords::FAILURE_VARIABLE . ');
+                    } else {
+                        $logger->error(' . ReservedKeywords::FAILURE_VARIABLE . ');
+                    }';
+
                 break;
 
             default:
-
+                // something went terribly wrong ...
+                throw new GeneratorException(
+                    sprintf(
+                        'Unknown enforcement type "%s", please check configuration value "enforcement/processing" and %s annotations within %s',
+                        $type,
+                        Processing::ANNOTATION,
+                        $structureName
+                    )
+                );
                 break;
         }
 
         return $code;
+    }
+
+    /**
+     * Will inject enforcement processing for a certain function.
+     * Will take default processing code into account and check for custom processing configurations
+     *
+     * @param string             $bucketData         Payload of the currently filtered bucket
+     * @param string             $structureName      The name of the structure for which we create the enforcement code
+     * @param string             $preconditionCode   Default precondition processing code
+     * @param string             $postconditionCode  Default post-condition processing code
+     * @param FunctionDefinition $functionDefinition Function definition to create the code for
+     *
+     * @return null
+     */
+    protected function injectFunctionEnforcement(
+        & $bucketData,
+        $structureName,
+        $preconditionCode,
+        $postconditionCode,
+        FunctionDefinition $functionDefinition
+    ) {
+        $functionName = $functionDefinition->getName();
+
+        // try to find a local enforcement processing configuration, if we find something we have to
+        // create new enforcement code based on that information
+        $localType = $this->filterLocalProcessing($functionDefinition->getDocBlock());
+        if ($localType !== false) {
+            // we found something, make a backup of default enforcement and generate the new code
+            $preconditionCode = $this->generateCode($structureName, 'precondition', $localType);
+            $postconditionCode = $this->generateCode($structureName, 'postcondition', $localType);
+        }
+
+        // Insert the code for the static processing placeholders
+        $bucketData = str_replace(
+            array(
+                Placeholders::ENFORCEMENT . $functionName . 'precondition' . Placeholders::PLACEHOLDER_CLOSE,
+                Placeholders::ENFORCEMENT . $functionName . 'postcondition' . Placeholders::PLACEHOLDER_CLOSE
+            ),
+            array($preconditionCode, $postconditionCode),
+            $bucketData
+        );
     }
 }