+++ src/StreamFilters/PreconditionFilter.php
@@ -87,8 +87,7 @@ class PreconditionFilter extends AbstractFilter
 
                     } else {
                         // Get the code for the assertions
-                        $code = $this->generateCode($functionDefinition->getAllPreconditions());
-
+                        $code = $this->generateCode($functionDefinition->getAllPreconditions(), $functionName);
 
                         // Insert the code
                         $bucket->data = str_replace(
@@ -117,16 +116,18 @@ class PreconditionFilter extends AbstractFilter
      * Will generate the code needed to enforce made precondition assertions
      *
      * @param \AppserverIo\Doppelgaenger\Entities\Lists\TypedListList $assertionLists List of assertion lists
+     * @param string                                                  $functionName   The name of the function for which we create the enforcement code
      *
      * @return string
      */
-    private function generateCode(TypedListList $assertionLists)
+    protected function generateCode(TypedListList $assertionLists, $functionName)
     {
         // We only use contracting if we're not inside another contract already
         $code = '/* BEGIN OF PRECONDITION ENFORCEMENT */
         if (' . ReservedKeywords::CONTRACT_CONTEXT . ') {
             ' . ReservedKeywords::PASSED_ASSERTION_FLAG . ' = false;' .
-            ReservedKeywords::FAILURE_VARIABLE . ' = array();';
+            ReservedKeywords::FAILURE_VARIABLE . ' = array();' .
+            ReservedKeywords::UNWRAPPED_FAILURE_VARIABLE . ' = array();';
 
         // We need a counter to check how much conditions we got
         $conditionCounter = 0;
@@ -155,7 +156,7 @@ class PreconditionFilter extends AbstractFilter
             }
 
             // close the or-combined wrap
-            $code .= 'if (empty(' . ReservedKeywords::FAILURE_VARIABLE . ')) {' .
+            $code .= 'if (empty(' . ReservedKeywords::FAILURE_VARIABLE . ') && empty(' . ReservedKeywords::UNWRAPPED_FAILURE_VARIABLE . ')) {' .
                 ReservedKeywords::PASSED_ASSERTION_FLAG . ' = true;
                 }}';
 
@@ -165,8 +166,7 @@ class PreconditionFilter extends AbstractFilter
 
         // Preconditions need or-ed conditions so we make sure only one condition list gets checked
         $code .= 'if (' . ReservedKeywords::PASSED_ASSERTION_FLAG . ' === false){' .
-            ReservedKeywords::FAILURE_VARIABLE . ' = implode(" and ", ' . ReservedKeywords::FAILURE_VARIABLE . ');' .
-            Placeholders::PROCESSING . 'precondition' . Placeholders::PLACEHOLDER_CLOSE . '
+            Placeholders::ENFORCEMENT . $functionName . 'precondition' . Placeholders::PLACEHOLDER_CLOSE . '
             }}
             /* END OF PRECONDITION ENFORCEMENT */';